third_party/blink/web_tests/fast/forms/select/option-relayout-modified-text-expected.txt

Modifying text content of an
changes the width of the .On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS refSelect.offsetWidth is > testSelect.offsetWidth
PASS testSelect.offsetWidth is refSelect.offsetWidth
PASS successfullyParsed is true

TEST COMPLETE


Pass if the two selects below have the same width.

OPTION TEXT


OPTION TEXT
